requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
boto3==1.21.32
mysql-connector-python==8.0.22
python-dotenv==0.21.0
requests
uuid
PyJWT

